pyproject.toml
[project]
name = "customs-trade-classification"
version = "1.0.0"
description = "Real-world customs trade classification RL environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "pydantic>=2.7.4",
    "openai>=1.40.0",
    "tenacity>=8.5.0",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "rapidfuzz>=3.0.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server", "data"]

[tool.setuptools.package-data]
"*" = ["*.*"]
